Update MidiWriter2 class for updated midi API
The MidiEvent class has an internal constructor.  Previously, while
building against the source of the Sanford midi library, I had modified
the constructor to be public.  Now that I'm using the nuget package, the
constructor is internal again.  In MidiWriter2, I modify the code that
was creating MidiEvents to return MidiMessages instead.