rewriting how commands are parsed, part 1
the code to parse commands is a mess.  this change tries to simplify by
getting rid of the map which didn't handle all cases and will use a case
statement instead.  unit tests will be written to verify the
implementation too.  the map approach might have worked in a language
like F# that supports currying, but because some commands have one
parameter and others take two, the single map approach became too
complicated.