replace default trace listener
This change replaces the default trace listener and disables displaying
an assert dialog when running tests.  The replacement listener throws an
exception when an assert fails.  This causes debug assert failures to
cause the running test to fail, and prevents an assert dialog from
blocking tests.  I don't know why this isn't the default for Visual
Studio.