parse '@' (empty) effect letter
When a command/effect column has no initial letter, it appears as a . in
ST/IT/OpenMPT etc but its letter value is the character before 'A',
which is '@'.  trackers interpret empty as a continuation of the
previous effect letter, but here it is interpreted as the None command
type.  It is up to the midi writer to keep track of the previous effect
and apply the new x and y values.  Note that for commands which take a
single value, such as Axx, the midi writer will need to rebuild the
double short value from x and y.  alternatively, the channel event class
could have another property which represents the combined xy value which
can be interpreted as xx as needed.