update TestNoteDelay1 test
TestNoteDelay1 was originally intended to test the SD1 effect, and uses
the sd.s3m file which contains 10 rows, the first 3 contain .00, SD1,
and SD2, which produce a triplet rhythm.  At some point, the test file
and test got out of sync, and this test was failing.  To update the
test, it is modified to match the test file and test accessing multiple
rows and ensuring that SD1 and SD2 are both read correctly.