drop midi channels above 16
Because the program works by mapping samples/instruments in ScreamTracker to MIDI channels, songs that have 16 or more samples produce midi channels that are out of range.  In this case, print a message and drop them.  Future enhancements could support a mapping between multiple samples to a single channel but the mapping would have to be user provided.