support mapping pattern channels to midi channels
For songs that use more than one sample or instrument per channel, you can end up with a score with lots and lots of tracks and it can be hard to follow a particular part.  there's now an option to map each channel to its own midi channel regardless of which sample is playing on that tracker channel.
in addition, when mapping samples to channels, the channel number is decreased by one so that 16 channel songs do not go above the midi max channel limit because the MIDI library being used stores channel 1 as 0.